move next_unsealed_results to a separate file
